Remove bad "" operators
Summary:
Feature: Basic Code Fixes

In C++11, "" is not parsed as before.  So "A""B" is not the same as "AB".

Instead, whitespace is required, like: "A" "B"

This adds that whitespace everywhere gcc-4.7 found it in our builds.

Test Plan:
Built with gcc-4.7.
'git show -w' shows zero changes.

Reviewers: flamingcow, pivanof, jeremycole, andrew-ford

Reviewed By: pivanof

CC: jtolmer, MarkCallaghan

Differential Revision: https://reviews.facebook.net/D14895