Split mysqld--help-notwin test into tests with and without profiling
Summary:
Feature: Basic Test Fixes

If we build with perfschema, this test can't work, since we disable
profiling for release builds, but not debug builds, and thus have
different command-line parameters for each build.

This splits this test so it can actually work, when it's actually run,
by splitting it into profiling and non-profiling versions.

This diff also adds the ability to support a 'have_noprofiling" test.

Test Plan: When perfschema is enabled, these tests pass.

Reviewers: flamingcow, pivanof, jeremycole, andrew-ford

Reviewed By: pivanof

CC: jtolmer, MarkCallaghan

Differential Revision: https://reviews.facebook.net/D14955