Add public phabricator config and testing support for WebScaleSQL
Summary:
This will let us handle reviews with external groups directly.

This also causes a "postponed" test to appear in the public diff page,
on each revision created with 'arc diff'.

This doesn't actually launch any tests, but it creates the "postponed"
test which allows for our internal testing sytem (and perhaps others) to
be pointed at this page, and fill out the test results from a test run.

Also added the created cache file to .gitignore.

Test Plan:
This is a combination of two, tested, diffs:

This diff was reviewed via public phabricator, so that works:
	https://reviews.facebook.net/D15621

And this diff was tested by using it to test a follow-up diff:
	https://reviews.facebook.net/D16713

For reference, the follow-up diff in question is here:
	https://reviews.facebook.net/D16629

Reviewers: pivanof, inaam-rana

Reviewed By: pivanof

CC: jtolmer

Differential Revision: https://reviews.facebook.net/D15621
Differential Revision: https://reviews.facebook.net/D16713