Use single quotes for perl paths, in case of special symbols
Summary:
Feature: Basic Test Fixes

Jenkins sometimes puts an @ in the directory name.

With this value in "s, in perl, that is interpreted as a list, and fails.

With it in 's, it works fine.

Test Plan:
Changed my 5.6 dir to 5@6, and mtr failed.
With this change, it now passes.

Reviewers: flamingcow, pivanof, jeremycole, andrew-ford, inaam-rana, liang.guo.752

Reviewed By: pivanof

CC: jtolmer, MarkCallaghan

Differential Revision: https://reviews.facebook.net/D15945