Adjust linking flags to not embed some libs
Summary:
Feature: Basic Build Fixes

Adjust mysql linking flags to not embed openssl and zlib in static libs.

By default, libmysqlclient.a includes a copy of the openssl and
zlib it was linked against, thanks to a MERGE_LIBRARY call.  This isn't
ideal in fbcode's third-party environment where we want targets to have
to link in openssl on their own.

This change removes the ssl and zlib libraries from the merge and
instead sets them as a dependency for the build, so that other built
targets, such as the command line tools, still end up properly linked
but libmysqlclient.a doesn't include embedded copies.  It also removes
the munging of the library names when doing a static build so that they
are not called "imported_crypto" and "imported_openssl" and instead are
called by their normal names.

Test Plan:
mtr; also builting in third-party, confirming the .a file
does not contain openssl symbols and that building a third-party library
that depends on libmysqlclient works as expected

Reviewers: flamingcow, pivanof, jeremycole, andrew-ford, inaam-rana

Reviewed By: pivanof

CC: jtolmer, MarkCallaghan

Differential Revision: https://reviews.facebook.net/D15657