Protect against writing to nullptr in net_write_buff
Summary:
Feature: Basic Bug Fixes

There are reports of segfaults from mysql trying to write when
there is no valid write buffer (typically after an aborted connect,
apparently).  This protects against writing to net->write_buf when it
isn't valid.

It's not fully clear if this is a 5.6 issue or an issue with the async
patch.  The codepaths using it seem to trigger only in very rare cases
and are inside of conenction pooling code that retries connections using
the same MYSQL object, which is arguably invalid to begin with.

Test Plan: mtr

Reviewers: pivanof, inaam-rana

Reviewed By: pivanof

CC: jtolmer, MarkCallaghan, flamingcow, jeremycole, andrew-ford, liang.guo.752, chip

Differential Revision: https://reviews.facebook.net/D17013