Page cleaner should do LRU flushing regardless of server activity
Summary:
WebScaleSQL Feature: Flushing Improvements

This is for upstream bug#70500 and bug#71988
page_cleaner should work whether or not there is server activity.
Its iterations become a noop when there is no work to do but we
should not tie it to the server activity.

The page_cleaner thread does spurious background flushing
because of conditional sleep between iterations. The solution
is not to make sleep dependent on server activity etc.

Test Plan: mtr

Reviewers: steaphan, pivanof

Reviewed By: steaphan

CC: MarkCallaghan, jtolmer, jeremycole, flamingcow, andrew-ford, pengt

Differential Revision: https://reviews.facebook.net/D16617