Avoid unnecessary buf_flush_list() in startup code path
Summary:
Feature: Basic Performance Fixes

This is a performance regression bug in upstream code. Opened
Bug#70899 to track this.

In InnoDB startup code innobase_start_or_create_for_mysql()
after creating rsegs we flush the entire buffer pool. The
intent is to force trx_sys page to the disk. We can reach
this code path after doing recovery. In this case we can
potentially have millions of dirty pages in the buffer pool.
This can seriously increase the recovery time.

The fix is trivial. Check if we are coming from recovery code
path and don't flush buffer pool in that case because during
recovery we don't create rsegs.

Test Plan: run mtr

Reviewers: pivanof, steaphan, liang.guo.752

Reviewed By: steaphan

CC: jtolmer, MarkCallaghan, flamingcow, jeremycole, andrew-ford, pengt

Differential Revision: https://reviews.facebook.net/D16623