Share more structures
Summary:
Feature: Basic Code Changes

This just moves some structure definitions from inside a
single .cc file to a shared .h file, with a few tweaks to
allow these structures to be shared.

On its own, it should have no actual effect.  It is a non-change.

This is a base change required for lots of stuff,
including innodb_fake_changes, stats, etc....

I am factoring this out to allow diff re-ordering.

Test Plan:
Build is clean, and so are mtr tests.

This code was factored out of the existing basic stats diff.
There are no overall code changes once that diff is applied.

This change has been running in production for about a year now.
All is still well, so this is pretty safe.

Reviewers: pivanof, inaam-rana, liang.guo.752

Reviewed By: pivanof

CC: jtolmer, MarkCallaghan, flamingcow, jeremycole, andrew-ford

Differential Revision: https://reviews.facebook.net/D16203