Add JUnit reporting support to the MySQL test suite runner.
Summary:
Feature: JUnit Support for MTR

Produces JUnit test reports that can be consumed by tools such as Jenkins.

Two new options are added to MTR:

  --junit-output=FILE   Output JUnit test summary XML to FILE.
  --junit-package=NAME  Set the JUnit package name to NAME for this test run.

A test run using junit reporting would look like:

  ./mtr --comment=rpl_row --junit-package=rpl_row --junit-output=rpl_row.xml \
        --suite=rpl --mysqld=--binlog-format=row

Typically, the package name should be the name for the test run, in
this case the suite name and its variation. Tests results are written
to the rpl_row.xml file (including test timing and output). If using
with Jenkins, the XML file can be used for publishing test results and
Jenkins would be able to show how many test failed, test duration and
whether to mark the build as failed or unstable (if there were failing
tests that succeeded on retry).

Test Plan: mtr

Reviewers: steaphan

Reviewed By: steaphan

Differential Revision: https://reviews.facebook.net/D17139