Fix mysqld--help-notwin-profiling output for 5.6.17
Summary:
Feature: Basic Test Fixes

Another command-line option was added in MySQL-5.6.17, but this version
of this test is new in WebScaleSQL, so it wasn't updated.

Since I only ran all the tests with the default builds, I did so with
PerfSchema disabled, and so this test was skipped, and this change
was missed.  This diff just fixes that one .result file.

This will be squashed with: https://reviews.facebook.net/D15753

Test Plan:
Built with perfschema enabled, with profiling, ran:
mtr --record main.mysqld--help-notwin-profiling

Built with perfschema enabled, without profiling, ran:
mtr --record main.mysqld--help-notwin

Reviewers: pivanof, inaam-rana, darnaut

Reviewed By: pivanof

CC: jtolmer

Differential Revision: https://reviews.facebook.net/D17529