Fix seg fault in new code path for reduced doublewrite mode recovery
Summary:
Feature: Reduced Double-Write Mode

A new code-path in 5.6.16 for fixing certain page 0 corruptions (which were
all completely broken in 5.6.12, where this feature was first implemented)
did not handle the "page" being null, which it will be if the server was in
this new reduced doublewrite mode when it crashed.

Fixed that.

I will squash this diff with this main diff in future releases:
https://reviews.facebook.net/D18939

Test Plan:
Full 'arc unit' test set shows this fix is clean.

Our Jenkins nightly tests caught this rare case, and this failure went
away after we pushed this fix internally.

Reviewers: pivanof, inaam-rana

Reviewed By: inaam-rana

Subscribers: jtolmer, WebScaleSQL

Differential Revision: https://reviews.facebook.net/D21435