Fix mtr failures related to processes exiting with 1
Summary:
WebScaleSQL Feature: Basic Test Fixes

We were seing various failures in the mtr tests related to processes exiting
with exit 1 but where there wasn't anything in the logs indicating anything
went wrong. These failures manifested as:

* mysqltest failed but provided no output
* Could not execute 'check-warnings' for test case ‘xxx.xxxxxxx' (res: 1):
* Valgrind failures where the valgrind report parsed out of the log contains 0 errors
* Server [mysqld.N - pid: NNNNN, winpid: NNNNN, exit: 256] failed during test run

We believe the cause of the problem was a race in my_safe_process related
to the handling of SIGCHLD. Our solution was to remove handling that signal
from the code, as it was redundant with other code in safe_process.cc that
checks for the child process terminating.

Test Plan:
Pushing this change caused all these failures to go away, across all our
test runs.  It continued to occur in other branches that did not have
this fix, until it was cherry-pick into those branches, then they all
went away there too.

Reviewers: pivanof, darnaut, inaam-rana

Reviewed By: inaam-rana

Subscribers: jtolmer, CalvinSun, WebScaleSQL

Differential Revision: https://reviews.facebook.net/D23127