Bug#19265668 CANT'T SHUTDOWN SERVER WITH --INNODB-FORCE-RECOVERY=6
When starting the server with innodb_force_recovery=6 the server would
hang if we try to shut it down. The root cause is that server is in
read-only mode with innodb_force_recovery=6 and the log_sys->lsn is
12 bytes(LOG_BLOCK_HDR_SIZE) after log_sys->last_checkpoint_lsn and
these 2 will never change. We have to identify this case.

rb#6280, approved by Jimmy.