Bug#17276183:5.6 SLAVE FAILS TO CORRECTLY READ BINLOG V3 EVENTS
Problem:
========
Master server is 4.1
Slave is 5.6.  Slave breaks in two ways,
1.It writes incorrect binlog filename to master.info
"test-bin.00" instead of "test-bin.000001"
2.Queries are truncated by about 4 bytes when they are read
from master

Analysis:
========
In old master checksums are not present, where as on
mysql-5.6 slave checksums are enabled and by default
"slave_sql_verify_checksum=1". Each event's checksum is
stored as part of last 4 bytes of that event. Hence when
replication is happening between 4.1 and 5.6 the new slave
interprets the last four bytes of events as checksum and
thinks the actual event to be contained within event_len-4.
Because of this each event is getting truncated by 4 bytes
on slave. For example when IO thread receives the first
rotate event the binlog name is stored at the end of the
event as master-bin.000001. When read_log_event code tries
to read the event it assumes checksum is available within the
event and converts actual event's length to
event_len = event_len - 4. Because of this binlog name
becomes master-bin.00 and gets stored in master.info
incorrectly. When server restarts the above incorrect name
is considered as binlog name and IO thread fails to start.

Secondly START_LOG_EVENT_V3 events are sent only once when
the server starts for the first time. Within the code
checksum_alg is set to "BINLOG_CHECKSUM_ALG_OFF" upon
receiving the Start_log_event_v3 event.The same checksum_alg
is used for all the following events. This makes replication
to work fine as long as server is not restarted.  If server
is restarted since there will be no more start event
events will get truncated by 4 bytes.

One solution for this problem is to start slave server
with slave_sql_verify_checksum=0. This will solve both the
above mentioned problems.

Hence as part of fix when new slave connects to a old master
with binlog version less than 4 a warning is written to the
error log which expects users to restart their slaves with
"slave_sql_verify_checksum=0".