Bug #18817867 MYSQLDUMP DOES NOT REPORT ERROR 28 WHEN               WRITING TO FULL NFS MOUNT
This is a follow up patch to this bug.
my_sync() which is added as part of the fix in mysqldump.c
is failing on MacOS when we try to dump the output to stdout.
On MacOS it will give ENOTSUP(errno 45) and in other OS we
will see EBADF. So, we can ignore this errno.