Bug#19545298 	REPLICATION FAILS WITH ERRNO 1755
Sometimes the slave IO thread will left a partial group in current
relay log when it is killed or stopped. After it is started, the
new relay log is created on rotation, a pair of ROTATE_EVENT and
FORMAT_DESCRIPTION_EVENT is replicated from master and written
into the new relay log firstly, if auto position is disabled, then
the remaining part of group is replicated, but currently we let
the worker roll back the current partial group when applying the
ROTATE_EVENT with MTS enabled, then the slave will encounter
some problems when applying the remaining part of group in the
new relay log, ERRNO 1755 is one problem of them.

To fix the problem, we do not let the worker roll back the partial
group if auto position is disabled. If auto position is enabled,
we let the worker roll back the partial group and gracefully finish
its work, before starting to apply the new (complete) copy of the
group.