Fix a bug where a deprecation warning prevents mysqld from starting
Summary:
WebScaleSQL Feature: Basic Code Fixes

This is to fix a bug from an upstream commit "Bug#19507567:RENAME VARIABLE
BINLOGGING_IMPOSSIBLE_MODE", in which the deprecated warning of
binlogging_impossible_mode prevents mysqld to start.

This is the upstream commit, and the line after which the "break;" should have been:

https://github.com/mysql/mysql-server/commit/e41df51f634a14e740d06a6bf07bbd3e1232ebd9#diff-e7d51f293adc9257284fbdec858056dfR8408

It's pretty clearly a mistake, since it's supposed to only be adding a warning about it being deprecated.

Test Plan: Started mysqld successfully after the fix.

Reviewers: inaam-rana, jeremycole, pivanof

Reviewed By: pivanof

Subscribers: jtolmer, steaphan

Differential Revision: https://reviews.facebook.net/D30783