Port the Async MySQL client to 5.6
Summary:
Feature: Async Client

This diff builds loosely upon the 5.1 client work and brings async to
5.6.  It is not fully feature complete (no SSL. no compression, only
supports the default native auth) but it is sufficient for our needs and
we can extend it.

This breaks the API slightly by cleaning it up and making it more
succinct.  Code using the old async API will need to change in minor
ways (though a compatibility layer could be introduced).

This also adds a #define to indicate the nonblocking client is available.

This changes slightly the ABI and so it also renames 'libmysqlclient' to
'libwebscalesqlclient' -- however, this is compatible with almost every
use case (changes are additive only, but do include changes to some data
structures).

Test Plan: mtr, testing against internal tools

Reviewers: chip, inaam-rana, pivanof, darnaut, jeremycole

Reviewed By: jeremycole

Subscribers: fe, darnaut, methane, steaphan, atcurtis, chip, liang.guo.752, andrew-ford, jeremycole, flamingcow, MarkCallaghan, jtolmer

Differential Revision: https://reviews.facebook.net/D17031