Fix occasional test failure caused by float-to-truncated-string
Summary:
Feature: Basic Test Fixes

This is to fix the test flakiness in main.ctype_binary, where a table is
created via concat(rand()). The derived varchar column has size 23, and
sometimes rand() will give a number which requries many fractional digits
to be represented in decimal, causing value errors in string truncation.

I give the rand() function a seed 0 so the output becomes deterministic
and the number is within the column size.

Test Plan:
This test fails only rarely, when it returns a very small fractional
value, which must be represented with many leading zeros in decimal.

This change makes this value deterministic instead of pseudo-random.

Reviewers: inaam-rana, pivanof, weixiang.zhai, plinux, xiaobin.lin, darnaut, jeremycole

Reviewed By: jeremycole

Subscribers: tian.xia

Differential Revision: https://reviews.facebook.net/D31119