When using MRG_MyISAM , and have the cascade slaves, the third slave may crash.
Summary:
WebScaleSQL Feature: Basic Bug Fixes

Suppose there are 4 servers in the environment. Row-based binlog format are
used.
M2 is the slave of M1;
M3 is the slave of M2;
M4 is the slave of M3;

Then we create MRG_MyISAM table, and insert one row.

It is because the RBR requires record the table_map event before Rows_wirte event.
And the table_map event will list all the related tables , (t, t1, t2) in this case.

In sql_thread of M3, (t, t1, t2) is expanded to be ((t, t1, t2), t1, t2).
In M4, when replays this event, the rli.tables_to_lock will be the link as t->t1->t2->t1->t2
But because the first t1 is in part of  MRG_MyISAM, its m_tabledef is not initialized. This causes the M4 crash when testing compatibility (table_def::compatible_with).

Ref: http://bugs.mysql.com/bug.php?id=69574

Test Plan: ./mtr main.bug69574 passed.

Reviewers: pengt, santoshb, steaphan, jeremycole

Reviewed By: jeremycole

Differential Revision: https://reviews.facebook.net/D31335