Fix a few type errors in the last push
Summary:
Feature: Logical Read-Ahead

There were a few minor type errors in an earlier diff.  A "return TRUE"
was included in a void function, and io_submit returns "int", but it was
being stored and compared in an "unsigned long".

This should have been part of:
https://reviews.facebook.net/D23463

...and I will squash this with that diff in future branches.

Test Plan:
Now it builds, it did not before.
I will run the full test suite on this diff.

Reviewers: weixiang.zhai

Reviewed By: weixiang.zhai

Differential Revision: https://reviews.facebook.net/D32379