Add rpl_read_size option
Summary:
WebScaleSQL Feature: rpl_read_size

This adds the my.cnf variable "rpl_read_size" with a default
value of 8kb. It must be a multiple of 4kb.

This sets the size for reads done from the relay log and binlog.
Making it larger might help with IO stalls while reading these files
when they are not in the OS buffer cache.

Test Plan:
Added mtr test cases

This has been used in production at Facebook for > 1 year.

Reviewers: pivanof, darnaut, inaam-rana, xiaobin.lin, santoshb, jeremycole, weixiang.zhai

Reviewed By: jeremycole

Subscribers: CalvinSun, jtolmer, steaphan

Differential Revision: https://reviews.facebook.net/D23493