Remove another extra file from upstream release tarball
Summary:
WebScaleSQL Feature: Basic Build Fixes

Docs/INFO_SRC is re-generated every build, so need not be in the repo.

This problem was reported externally as:
* https://github.com/webscalesql/webscalesql-5.6/issues/15

This will be merged with the main file-removal diff in future branches:
* https://reviews.facebook.net/D15627

Test Plan:
Built after removing this file, and it built fine.

Also confirmed it creates this file if I build in-dir.

Reviewers: weixiang.zhai, darnaut, jeremycole, pivanof, inaam-rana

Reviewed By: inaam-rana

Subscribers: fe

Differential Revision: https://reviews.facebook.net/D33609