Full support for headerless compressed InnoDB pages
Summary:
WebScaleSQL Feature: 2012 InnoDB Compression

This exposes more zlib config parameters as system vars. The most
important of these variables is innodb_zlib_wrap which makes zlib not
write adler32 checksums between compressed blocks. The default value for
wrapping is OFF now. This should save some calls to adler32 and some
space. Changes are backwards compatible. In other words old compressed
server data can be read correctly with a server that has this change.

This is because the transaction log records are updated to include the
value for wrap and the compression strategy in addition to compression
level. This was fairly straightforward because the same was done earlier
with compression level. Prior to this change, we stored the compression
level in a byte. Now, the spare bits in that byte are used to store
values for the new options.

innodb_zlib_wrap and innodb_zlib_strategy also effect blob compression.

It's ok to change these values dynamically - it won't break xtrabackup
or crash recovery.

This diff includes the full set of tests that have been accumulated over
the years to continue to confirm this keeps working as we change things
further.

Also, this changes innodb_zlib_wrap randomly in the main stress tests, to
stress test this change as well.

Test Plan:
mtr - with all these new tests, with and without PerfSchema.  All passed.

Plus, this has been in production at Facebook for years.

Reviewers: inaam-rana, pivanof

Reviewed By: pivanof

Subscribers: WebScaleSQL, jtolmer

Differential Revision: https://reviews.facebook.net/D17721