Fix unit test in server-side statement timeout diff
Summary:
WebScaleSQL Feature: Server-Side Statement Timeout

It looks like the unit test included in the orignial diff was broken.
It seems to have included stale code, and did not build.

Our test system missed this, because it doesn't build gunit tests.

This will be squashed with the original diff in future branches:
* https://reviews.facebook.net/D22881

Test Plan:
Built locally on a system with gunit tests enabled.

It would be worth actually running this test to, to be sure it still
passes... I did not do that.

Reviewers: pivanof, darnaut, jeremycole, inaam-rana

Reviewed By: inaam-rana

Subscribers: jtolmer, CalvinSun

Differential Revision: https://reviews.facebook.net/D33687