Fix DTrace funtions in Async Client
Summary:
WebScaleSQL Feature: Async Client

This fixes a bad code duplication with the upstream 6.0 DTrace patches,
which is in the WebScaleSQL Async Client patch.

This code is not actully built (the macro is defined to empty), when not
built with dtrace (or systemtap) so we did not detect this problem.

Fortunately, someone else did, and reported it:
https://github.com/webscalesql/webscalesql-5.6/issues/17

I will squash this commit with the Async Client diff in future branches:
* https://reviews.facebook.net/D17031

Test Plan:
Installed systemtap on a dev machine, and confirmed the error.

Then, I fixed the incorect argument to the macro (based on the other
similar function), and built again, and all is now well.

Reviewers: chip, pivanof, darnaut, jeremycole, inaam-rana

Reviewed By: inaam-rana

Subscribers: liang.guo.752, chip, atcurtis, steaphan, methane, darnaut, fe

Differential Revision: https://reviews.facebook.net/D33777