Attempt to stabilize innodb.innodb_doublewrite*
Summary:
There appear to be 3 failure modes in this test currently. This fix is
primarily aimed at fixing the most common, random.choice raising
IndexError. The cause of that intermittent problem is that there aren't
always pages in the doublewrite buffer from spaces other than 0. Rather
than continue to figure out how to manipulate all the background processing
in InnoDB that does page flushing, change the tactic of the test. Instead of
relying on there being existing pages for the target space_id in the doublewrite
buffer to manipulate, write a page to the doublewrite buffer while the
server is shutdown.

Test Plan: ./mtr --valgrind --parallel=16 --repeat=16 ...

Reviewers: inaam-rana

Reviewed By: inaam-rana

Differential Revision: https://reviews.facebook.net/D36945