Fixes for main.connection_timeout
Summary:
The test has never worked when mtr is run with --embedded-server
or --ps-protocol.

Disable the test for embedded because that config doesn't support
creating additional connections.

For the prepared statement protocol, I had to make a code change
so that the correct error message stopped getting swallowed due to
upstream commit https://github.com/mysql/mysql-server/commit/3b3983a4
which made it so that 2013 was always returned.

Squash with D17019

Test Plan:
  ./mtr main.connection_timeout
  ./mtr --embedded-server main.connection_timeout
  ./mtr --ps-protocol main.connection_timeout
  ./mtr --force --timer --big-test --testcase-timeout=90 \
    --parallel=auto --experimental=collections/default.experimental \
    --ps-protocol --skip-test-list=collections/disabled-weekly.list

Reviewers: darnaut, inaam-rana, pivanof

Reviewed By: pivanof

Subscribers: kastauyra, chip

Differential Revision: https://reviews.facebook.net/D42993