Fixed MTR failure main.events_1.test
Summary:
The failure was caused by that the specified end date 20151010 when creating a event
actually was expired. Fixed this issue by changing it to 20351010 to make sure
it won't get expired in 20 years.

The bug has been fixed in Facebook's branch https://reviews.facebook.net/D48831

Test Plan: MTR

Reviewers: inaam-rana, xiaobin.lin, santoshb, jeremycole

Reviewed By: jeremycole

Differential Revision: https://reviews.facebook.net/D49005